Refactored tensor.py into package (#1305)
* refactored tensor.py to tensor folder by spliting into classes

* fixed name of __init__
