+++ syft/tensor/__init__.py
@@ -1,0 +1,4 @@
+#Imports of tensors for ease of reference
+from base_tensor import BaseTensor
+from int_tensor import IntTensor
+from float_tensor import FloatTensor
